+++ .github/jobs/test-used-linker.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -xeuo pipefail
+
+RUSTDOCFLAGS='--cfg linktime_used_linker' \
+    RUSTFLAGS='-D warnings --cfg linktime_used_linker' \
+    cargo test --target $TARGET