+++ .github/jobs/test.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 set -xeuo pipefail
 
-cargo test --target $TARGET
+RUSTFLAGS='-D warnings' cargo test --target $TARGET