+++ ctor/examples/ctor-basic.rs
@@ -1,4 +1,5 @@
 //! Basic example of using the `ctor` crate.
+#![cfg_attr(linktime_used_linker, feature(used_with_arg))]
 
 use ctor::ctor;