+++ ctor/src/lib.rs
@@ -1,10 +1,15 @@
-#![cfg_attr(feature = "used_linker", feature(used_with_arg))]
 #![recursion_limit = "256"]
 #![no_std]
 #![doc = include_str!("../docs/BUILD.md")]
 //! # ctor
 #![doc = include_str!("../docs/PREAMBLE.md")]
 #![doc = include_str!("../docs/GENERATED.md")]
+// Used as part of ctor collection
+#![cfg_attr(
+    all(target_vendor = "apple", linktime_used_linker),
+    feature(used_with_arg)
+)]
+#![cfg_attr(linktime_used_linker, doc(test(attr(feature(used_with_arg)))))]
 
 #[cfg(feature = "std")]
 extern crate std;
@@ -230,7 +235,6 @@ pub mod declarative {
 /// Print a startup message (using `libc_print` for safety):
 ///
 /// ```rust
-/// # #![cfg_attr(feature="used_linker", feature(used_with_arg))]
 /// # #[cfg(not(miri))] mod test {
 /// # use ctor::ctor;
 /// use libc_print::std_name::println;
@@ -249,7 +253,6 @@ pub mod declarative {
 /// Make changes to `static` variables:
 ///
 /// ```rust
-/// # #![cfg_attr(feature="used_linker", feature(used_with_arg))]
 /// # mod test {
 /// use ctor::*;
 /// use std::sync::atomic::{AtomicBool, Ordering};
@@ -265,7 +268,6 @@ pub mod declarative {
 /// Initialize a `HashMap` at startup time:
 ///
 /// ```rust
-/// # #![cfg_attr(feature="used_linker", feature(used_with_arg))]
 /// # mod test {
 /// # use std::collections::HashMap;
 /// # use ctor::*;
@@ -290,7 +292,6 @@ pub mod declarative {
 /// is run at startup time.
 ///
 /// ```rust
-/// # #![cfg_attr(feature="used_linker", feature(used_with_arg))]
 /// # mod test {
 /// # use ctor::*;
 /// #[ctor(unsafe)]
@@ -473,19 +474,6 @@ __declare_features!(
     naked {
         attr: [(naked) => (naked)];
     };
-    no_fail_on_missing_unsafe {
-        /// attr
-        /// Marks a ctor as unsafe. Required.
-        ///
-        /// The `ctor` crate will warn if there is no unsafe flag in the `ctor`
-        /// annotation. This warning for a missing unsafe keyword can be hidden
-        /// by passing `RUSTFLAGS="--cfg no_fail_on_missing_unsafe"` to Cargo.
-        attr: [(unsafe) => (no_fail_on_missing_unsafe)];
-        default {
-            (no_fail_on_missing_unsafe) => (no_fail_on_missing_unsafe),
-            _ => (),
-        }
-    };
     /// The priority of the constructor. Higher-`N`-priority constructors are
     /// run last. `N` must be between 0 and 999 inclusive for ordering
     /// guarantees (`N` >= 1000 ordering is platform-defined).
@@ -532,13 +520,41 @@ __declare_features!(
     std {
         feature: "std";
     };
-    used_linker {
-        /// crate
-        /// Applies `used(linker)` to all `ctor`-generated functions. Requires nightly and `feature(used_with_arg)`.
-        feature: "used_linker";
+    r#unsafe {
         /// attr
-        /// Mark generated functions for this `ctor` as `used(linker)`. Requires nightly and `feature(used_with_arg)`.
+        ///
+        /// Marks a ctor as unsafe. Required.
+        ///
+        /// The `ctor` crate will warn if there is no unsafe flag in the `ctor`
+        /// annotation. This warning for a missing unsafe keyword can be hidden
+        /// by passing `RUSTFLAGS="--cfg linktime_no_fail_on_missing_unsafe"` to
+        /// Cargo.
+        attr: [(unsafe) => (no_fail_on_missing_unsafe)];
+        default {
+            (linktime_no_fail_on_missing_unsafe) => (no_fail_on_missing_unsafe),
+            _ => (),
+        }
+    };
+    used_linker {
+        /// attr
+        ///
+        /// Mark generated functions pointers `used(linker)`. Requires nightly
+        /// for the nightly-only feature `feature(used_with_arg)` (see
+        /// <https://github.com/rust-lang/rust/issues/93798>).
+        ///
+        /// The can be made the default by using the `cfg` flag
+        /// `linktime_used_linker` (`RUSTFLAGS="--cfg linktime_used_linker"`).
+        ///
+        /// For a crate using this macro to function correctly with and without
+        /// this flag, it is recommended to add the following line to the top of
+        /// lib.rs in the crate root:
+        ///
+        /// `![cfg_attr(linktime_used_linker, feature(used_with_arg))]`
         attr: [(used(linker)) => (used_linker)];
+        default {
+            (linktime_used_linker) => used_linker,
+            _ => (),
+        }
     };
 );