+++ ctor/tests/expand-darwin/ctor_used_linker_attr.expanded.rs
@@ -23,3 +23,28 @@ fn foo() {
     };
     { __ctor_private_inner() }
 }
+#[allow(dead_code)]
+fn naked_foo() {
+    #[allow(unsafe_code)]
+    #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
+    fn __ctor_private_inner() {
+        {
+            ::std::io::_print(format_args!("foo\n"));
+        };
+    }
+    const _: () = {
+        #[allow(unsafe_code)]
+        #[link_section = "__DATA,__mod_init_func,mod_init_funcs"]
+        #[used(linker)]
+        static __CTOR_PRIVATE_REF: unsafe extern "C" fn() = {
+            #[allow(unused_unsafe)]
+            #[allow(unsafe_code)]
+            #[link_section = "__TEXT,__text_startup,regular,pure_instructions"]
+            extern "C" fn __ctor_private() {
+                { { __ctor_private_inner() } }
+            }
+            __ctor_private
+        };
+    };
+    { __ctor_private_inner() }
+}