+++ ctor/tests/expand-darwin/ctor_used_linker_attr.rs
@@ -4,3 +4,8 @@ use ctor::ctor;
 fn foo() {
     println!("foo");
 }
+
+#[ctor(unsafe, naked, used(linker))]
+fn naked_foo() {
+    println!("foo");
+}