+++ dtor/tests/expand-darwin/dtor_used_linker_attr.rs
@@ -1,0 +1,6 @@
+use dtor::dtor;
+
+#[dtor(unsafe, used(linker))]
+fn foo() {
+    println!("foo");
+}